I see now the data from one user only but this is not the way i want to retreive that data

+++ backend/src/main/java/pl/maciejklonicki/ytapp/users/UserRepository.java
@@ -3,7 +3,7 @@ package pl.maciejklonicki.ytapp.users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository <Users, Long> {
+public interface UserRepository extends JpaRepository <Users, Integer> {
 
     Optional<Users> findByEmail(String email);
 }
